:truck: relocate extensions.
