+++ commons/src/main/kotlin/tw/waterballsa/utopia/commons/extensions/PathExtensions.kt
@@ -1,4 +1,4 @@
-package tw.waterballsa.utopia.gaas.extensions
+package tw.waterballsa.utopia.commons.extensions
 
 import java.nio.file.Path
 import kotlin.io.path.createDirectories
@@ -6,13 +6,13 @@ import kotlin.io.path.createFile
 import kotlin.io.path.exists
 
 
-internal fun Path.createFileIfNotExists(): Path =
+fun Path.createFileIfNotExists(): Path =
     when (exists()) {
         true -> this
         false -> createFile()
     }
 
-internal fun Path.createFileWithFileName(fileName: String): Path =
+fun Path.createFileWithFileName(fileName: String): Path =
     createDirectories()
         .resolve(fileName)
         .createFileIfNotExists()
